rename main.py to app.py
