+++ src/power_grid_model_ds/visualizer.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: MPL-2.0
 
 try:
-    from power_grid_model_ds._core.visualizer.main import visualize
+    from power_grid_model_ds._core.visualizer.app import visualize
 except ImportError as error:
     raise ImportError(
         "Missing dependencies for visualizer: install with 'pip install power-grid-model-ds[visualizer]'"
