+++ tests/unit/visualizer/test_parsers.py
@@ -2,7 +2,6 @@
 #
 # SPDX-License-Identifier: MPL-2.0
 
-
 import numpy as np
 from numpy.typing import NDArray
